backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pytest
httpx
neo4j
psycopg2-binary
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
google-genai
python-dotenv
